pyproject.toml
[tool.poetry]
name = "advent-tree"
version = "0.1.0"
description = "Tree for Advent"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.7"
djangorestframework = "^3.14.0"
awsebcli = "^3.20.10"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
